Fix and enhance en/em dash translation
Depending on the back-end being used, '--' gets currently translated either to
an em dash (HTML, for instance) or an en dash (LaTeX). To allow convenient
usage of both types of dashes and resolve this inconsistency, translate (in
sync with LaTeX) '--' to en dash and '---' to em dash.

In order to avoid "breaking" existing documents, introduce an attribute named
'compat-mode' that is defined by default and must be undefined to enable the
en/em dash translation described above.

Signed-off-by: Christoph Jaeger <email@christophjaeger.info>